+++ backend/src/main/java/org/ukma/spring/crooodle/service/RoomService.java
@@ -5,13 +5,13 @@ import org.ukma.spring.crooodle.dto.*;
 import org.ukma.spring.crooodle.dto.common.PageResponseDto;
 
 public interface RoomService {
-    int create(int hotelId, RoomCreateRequestDto requestDto);
-    RoomAdminResponseDto getAdmin(int hotelId, int roomId);
-    void edit(int hotelId, int roomId, RoomEditRequestDto requestDto);
-    void delete(int hotelId, int roomId);
-    PageResponseDto<RoomAdminResponseDto> listAdmin(int hotelId, RoomCriteriaDto criteriaDto);
+    int create(Long hotelId, RoomCreateRequestDto requestDto);
+    RoomAdminResponseDto getAdmin(Long hotelId, Long roomId);
+    void edit(Long hotelId, Long roomId, RoomEditRequestDto requestDto);
+    void delete(Long hotelId, Long roomId);
+    PageResponseDto<RoomAdminResponseDto> listAdmin(Long hotelId, RoomCriteriaDto criteriaDto);
 
-    RoomResponseDto get(int hotelId, int roomId);
-
-    PageResponseDto<RoomResponseDto> list(int hotelId, @Valid RoomCriteriaDto criteriaDto);
+    RoomResponseDto get(Long hotelId, Long roomId);
+    PageResponseDto<RoomResponseDto> list(Long hotelId, @Valid RoomCriteriaDto criteriaDto);
 }
+
